update config structure and config reader
